Fix for index-related crash
Basically trying to drop an index that doesn't exist is a bad thing, but trying to add an index when one already exists is totally fine. Go figure.

This was previously necessary because some of the collections had an index that wasn't properly set up, but since the wipe happened we can be sure the indices are all set properly.